backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
psycopg[binary]==3.2.9
pydantic-settings==2.8.1
python-jose[cryptography]==3.3.0
